Fixed issues with manual picking, parameter guide and reanalysis
Before, there were many issues with manual picking of minis and clicking "Reanalyze" button in the parameter guide on a found mini, causing the data plot to get out of sync and the program generating errors. That's all fixed now. Also, the 10-90 rise time calculation is improved to be less affected by oscillation of the data near 10% and 90% due to noise.